refactor: web3-shared-constants (#5840)
